Started refactoring PoTDA to different classes
