+++ core/src/com/pottda/game/application/GameState.java
@@ -1,0 +1,16 @@
+package com.pottda.game.application;
+
+public enum GameState {
+    NONE,
+    RUNNING,
+    PAUSED,
+    OPTIONS,
+    MAIN_MENU,
+    MAIN_CHOOSE,
+    MAIN_CONTROLS,
+    WAITING_FOR_INVENTORY,
+    GAME_OVER,
+    RESTARTING;
+
+    static GameState gameState;
+}
