--- core/src/com/pottda/game/application/InventoryScreen.java
@@ -1,46 +1,0 @@
-package com.pottda.game.application;
-
-import com.badlogic.gdx.Game;
-import com.badlogic.gdx.Screen;
-
-/**
- * Screen that acts as top-level for the inventory sccreen
- */
-public class InventoryScreen implements Screen {
-
-
-    @Override
-    public void show() {
-
-    }
-
-    @Override
-    public void render(float delta) {
-
-    }
-
-    @Override
-    public void resize(int width, int height) {
-
-    }
-
-    @Override
-    public void pause() {
-
-    }
-
-    @Override
-    public void resume() {
-
-    }
-
-    @Override
-    public void hide() {
-
-    }
-
-    @Override
-    public void dispose() {
-
-    }
-}
