+++ core/src/com/pottda/game/model/Constants.java
@@ -1,0 +1,16 @@
+package com.pottda.game.model;
+
+public class Constants {
+
+    public static final String GAME_TITLE = "Panic on TDAncefloor";
+
+    public static final float SCALING = 2f;
+
+    public static final float WIDTH = 800;
+    public static final float HEIGHT = 480;
+    public static final float WIDTH_METERS = 30 * SCALING;
+    public static final float HEIGHT_METERS = 18 * SCALING;
+    public static final float HEIGHT_RATIO = WIDTH_METERS / WIDTH / SCALING;
+    public static final float WIDTH_RATIO = HEIGHT_METERS / HEIGHT / SCALING;
+
+}
